feat: Refactor login page to use Formik and Zod for validation, implement session handling and error notifications
chore: Remove middleware file and implement authentication directly in the new middleware

fix: Update axios interceptor to handle token and session expiration with improved error handling

feat: Add DeleteConfirmationModal component for RFID management

feat: Create RFIDAddModal and RFIDEditModal components for adding and editing RFID entries

feat: Implement RoomAddEditModal for managing room details with validation

feat: Add User management modals for editing and deleting users with confirmation

feat: Enhance UserDetailModal to fetch and display user details with image handling

fix: Update UserEditModal to handle image uploads and validations with improved error handling

chore: Reintroduce middleware for authentication with updated logic and matcher configuration
