--- src/app/middleware.ts
@@ -1,6 +1,0 @@
-export { auth as middleware } from "@/app/auth"
-
-
-export const config = {
-    matcher: ["/((?!api|_next/static|_next/image|favicon.ico|login).*)"],
-}
